Move API and Client to their own projects
