+++ StacyClouds.SwaAuth.Api/StacyClouds.SwaAuth.Api.csproj
@@ -1,0 +1,36 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+	  <Title>Static Web App API Authentication and Authorization</Title>
+	  <Version>0.1.1</Version>
+	  <Authors>Stacy Cashmore</Authors>
+	  <Company>StacyClouds</Company>
+	  <Description>A library of functions for .NET Azure Functions with Azure Static Web App authentication and authorization</Description>
+	  <PackageProjectUrl>https://github.com/StacyClouds/SwaAuth</PackageProjectUrl>
+	  <RepositoryUrl>https://github.com/StacyClouds/SwaAuth</RepositoryUrl>
+	  <PackageTags>swa; azure functions</PackageTags>
+	  <PackageOutputPath>..\pkg</PackageOutputPath>
+	  <PackageLicenseFile>LICENSE</PackageLicenseFile>
+	  <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
+	  <IncludeReferencedProjects>True</IncludeReferencedProjects>
+  </PropertyGroup>
+  
+   <ItemGroup>
+    <FrameworkReference Include="Microsoft.AspNetCore.App" />
+   </ItemGroup>
+  
+   <ItemGroup>
+     <None Include="..\LICENSE" Link="LICENSE">
+       <PackagePath>\</PackagePath>
+       <Pack>True</Pack>
+     </None>
+   </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\StacyClouds.SwaAuth.Models\StacyClouds.SwaAuth.Models.csproj" />
+  </ItemGroup>
+
+</Project>
