+++ StacyClouds.SwaAuth.Client/AuthenticationExtensions.cs
@@ -1,6 +1,5 @@
 ﻿using Microsoft.AspNetCore.Components.Authorization;
 using Microsoft.Extensions.DependencyInjection;
-using StaticWebAppAuthentication.Client;
 
 namespace StacyClouds.SwaAuth.Client;
