+++ StacyClouds.SwaAuth.Models/AuthenticationData.cs
@@ -1,0 +1,6 @@
+﻿namespace StacyClouds.SwaAuth.Models;
+
+public class AuthenticationData
+{
+    public ClientPrincipal? ClientPrincipal { get; set; }
+}
