+++ StacyClouds.SwaAuth.Models/ClientPrincipal.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace StacyClouds.SwaAuth.models;
+namespace StacyClouds.SwaAuth.Models;
 
 public class ClientPrincipal
 {
