--- StacyClouds.SwaAuth/models/AuthenticationData.cs
@@ -1,6 +1,0 @@
-﻿namespace StacyClouds.SwaAuth.models;
-
-internal class AuthenticationData
-{
-    public ClientPrincipal? ClientPrincipal { get; set; }
-}
